core: Define new internal dynamic receive buffering capability
Introduce a new capability called dynamic receive buffering.  This
is defined as an internal feature, specifically targeting the layering
of the rxm provider over tcp.  When dynamic rbuf is enabled, the tcp
provider will invoke a callback into rxm to obtain the receive
buffer that data will be copied into.  This avoids the double buffering
and copies that occur when running rxm over tcp, allowing tcp to
copy data directly into the application buffer.

To avoid needing to completely redesign the rxm provider and limit
the changes needed to the tcp provider, the general operation is as
follows:

RxM continues to post rx_buf's to the tcp provider.  However, the
buffers are reduced from the default eager buffer size (16k) to the
size of the base rxm header (struct rxm_pkt - 64 bytes).  The tcp provider
will read the first 64 bytes of each received message into the buffer
posted by rxm.  After this, it will invoke a callback to rxm to obtain
an iovec where the rest of the data will be copied.  The tcp provider
will then continue processing the receive, until it completes, then
generate an entry on the CQ.

In ideal cases, rxm will use the received header to locate the correct
application buffer where the data should be copied.  If no app buffer
is available, the message will be treated as unexpected and copied into
a temporary buffer.

Additional changes are needed to handle unexpected messages correctly
and internal rxm messages.

The use of this feature requires that rxm use shared receive contexts
(implementation restriction), and the SAR protocol is disabled.
SAR is no longer needed, and is replaced by an increased eager message
size.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>